start ISubmission service and move AvailableStatuses from SubmissionHelper to this class
