+++ tests/TASVideos.Core.Tests/Services/SubmissionServiceTests.cs
@@ -2,17 +2,16 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using TASVideos.Core.Services;
 using TASVideos.Data.Entity;
-using TASVideos.Data.Helpers;
 using static TASVideos.Data.Entity.SubmissionStatus;
 
-namespace TASVideos.Data.Tests.Helpers
+namespace TASVideos.Core.Tests.Services
 {
 	[TestClass]
-	public class AvailableStatusesTests
+	public class SubmissionServiceTests
 	{
-		private static DateTime OldEnoughToBeJudged
-			=> DateTime.UtcNow.AddHours(-1 - SiteGlobalConstants.MinimumHoursBeforeJudgment);
+		private readonly SubmissionService _submissionService;
 
 		private static DateTime TooNewToJudge => DateTime.UtcNow;
 
@@ -21,10 +20,18 @@ namespace TASVideos.Data.Tests.Helpers
 		private static readonly IEnumerable<PermissionTo> PublisherPerms = new[] { PermissionTo.SubmitMovies, PermissionTo.PublishMovies };
 		private static readonly IEnumerable<PermissionTo> Override = new[] { PermissionTo.OverrideSubmissionStatus };
 
+		public SubmissionServiceTests()
+		{
+			_submissionService = new SubmissionService();
+		}
+
+		private static DateTime OldEnoughToBeJudged
+			=> DateTime.UtcNow.AddHours(-1 - SiteGlobalConstants.MinimumHoursBeforeJudgment);
+
 		[TestMethod]
 		public void Published_CanNotChange()
 		{
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				Published,
 				Override,
 				OldEnoughToBeJudged,
@@ -49,7 +56,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Submitter_BasicPerms(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				BasicUserPerms,
 				OldEnoughToBeJudged,
@@ -77,7 +84,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Submitter_IsJudge(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				JudgePerms,
 				OldEnoughToBeJudged,
@@ -105,7 +112,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Submitter_IsPublisher(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				PublisherPerms,
 				OldEnoughToBeJudged,
@@ -133,7 +140,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Judge_ButNotSubmitter_BeforeAllowedJudgmentWindow(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				JudgePerms,
 				TooNewToJudge,
@@ -161,7 +168,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Judge_ButNotSubmitter_AfterAllowedJudgmentWindow(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				JudgePerms,
 				OldEnoughToBeJudged,
@@ -189,7 +196,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Publisher_ButNotSubmitter_BeforeAllowedJudgmentWindow_CanNotChangeStatus(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				PublisherPerms,
 				TooNewToJudge,
@@ -217,7 +224,7 @@ namespace TASVideos.Data.Tests.Helpers
 		public void Publisher_ButNotSubmitter_AfterAllowedJudgmentWindow(SubmissionStatus current, IEnumerable<SubmissionStatus> canChangeTo)
 		{
 			var expected = new[] { current }.Concat(canChangeTo).ToList();
-			var result = SubmissionHelper.AvailableStatuses(
+			var result = _submissionService.AvailableStatuses(
 				current,
 				PublisherPerms,
 				OldEnoughToBeJudged,
@@ -244,7 +251,7 @@ namespace TASVideos.Data.Tests.Helpers
 
 			foreach (var current in exceptPublished)
 			{
-				var result = SubmissionHelper.AvailableStatuses(
+				var result = _submissionService.AvailableStatuses(
 					current,
 					Override,
 					TooNewToJudge,
